Add default post visibility
