+++ core/context.py
@@ -5,7 +5,7 @@ def config_context(request):
     return {
         "config": Config.system,
         "config_identity": (
-            Config.load_identity(request.identity) if request.identity else None
+            request.identity.config_identity if request.identity else None
         ),
         "top_section": request.path.strip("/").split("/")[0],
     }
