+++ templates/activities/_menu.html
@@ -41,7 +41,7 @@
         {% csrf_token %}
         {{ form.text }}
         {{ form.content_warning }}
-        <input type="hidden" name="visibility" value="0">
+        <input type="hidden" name="visibility" value="{{ config_identity.default_post_visibility }}">
         <div class="buttons">
             <span class="button toggle" _="on click toggle .enabled then toggle .hidden on #id_content_warning">CW</span>
             <button>{% if config_identity.toot_mode %}Toot!{% else %}Post{% endif %}</button>
